mcp-langchain-test/requirements.txt
langchain-groq
langchain-mcp-adapters
langchain-community
langchain
langchain-openai
mcp
mcp[cli]
python-dotenv
langgraph
fastmcp